+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## [3.9.3] - 2026-05-21
+
+### Fixed
+
+- Fixed regex performance issue by converting optional-separator patterns to atomic-optional groups.
+
 ## [3.9.1] - 2026-05-18
 
 ### Fixed